+++ discovery_engine_core/core/src/stack.rs
@@ -32,7 +32,7 @@ use crate::{
 mod data;
 mod filters;
 mod ops;
-mod semantic_filtering;
+mod semantic_filters;
 
 pub use self::ops::Ops;
 pub(crate) use self::{
